Do not change type of _timeout for now.

+++ plugins/cache/memcached.py
@@ -38,7 +38,8 @@ DOCUMENTATION = '''
             section: defaults
       _timeout:
         default: 86400
-        type: float
+        type: integer
+        # TODO: determine whether it is OK to change to: type: float
         description: Expiration timeout in seconds for the cache plugin data. Set to 0 to never expire
         env:
           - name: ANSIBLE_CACHE_PLUGIN_TIMEOUT
